# Improve tile read speeds for some pyramid OME-TIFFs