Backported from a private PR.  This addresses the pyramid performance problem discussed with @sbesson.

Any time a SubIFD contains large amounts of non-pixel data (e.g. a full copy of the OME-XML), ```openBytes``` is slowed down substantially.  This changes the SubIFD reading in ```openBytes``` to only request a single SubIFD, and not even try to read the corresponding ```ImageDescription``` tag since it will be ignored anyway.

7cee004 is a minor improvement to reduce the number of tiles that are checked when reading pixel data.  It could be reverted if any objections.

I wouldn't expect this to address other issues with OME-TIFF read performance, but maybe points to IFD size instead of modality as a point for further investigation.

I would not expect tests to fail or memo files to be impacted.