@dgault 2022-02-16
The logic here makes sense to me and I don't see it having any negative consequences. Builds and tests have also remained passing for some time now and the performance of the repo tests has remained consistent.

I tried testing some tiled reads with the public sample files and generally speaking the times were either slightly quicker or more or less the same.

https://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/sub-resolutions/Fluorescence/

Without the PR:
Resolution 1: 16.656s
Resolution 2: 4.95s

With the PR
Resolution 1: 14.464s
Resolution 2: 4.201s

https://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/sub-resolutions/Brightfield/Leica-1/

Without the PR
Resolution 0: 19.827s
Resolution 1: 2.659s
Resolution 2: 0.772s

With the PR
Resolution 0: 15.466s
Resolution 1: 2.741s
Resolution 2: 0.829s

https://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/sub-resolutions/EM/

Without the PR
Resolution 1: 36.95s
Resolution 2: 7.041s

With the PR
Resolution 1: 29.681s
Resolution 2: 7.148s

https://downloads.openmicroscopy.org/images/OME-TIFF/2016-06/sub-resolutions/Z-stack/

Without the PR
Resolution 0: 26.783s
Resolution 1: 22.652s
Resolution 2: 20.32s

With the PR
Resolution 0: 25.933s
Resolution 1: 21.013s
Resolution 2: 19.685s

PR looks good from my perspective. Merging this for inclusion with 6.9.0